refactor(core): reduce the number of circular deps
This commit updates several import statements in the core package to decrease the number of
cycles detected by the dependency checker tool.
