+++ packages/core/src/render3/interfaces/type_checks.ts
@@ -6,9 +6,8 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {ComponentDef, DirectiveDef} from '..';
-
 import {LContainer, TYPE} from './container';
+import {ComponentDef, DirectiveDef} from './definition';
 import {TNode, TNodeFlags} from './node';
 import {RNode} from './renderer';
 import {FLAGS, LView, LViewFlags} from './view';
